Add better error handling to Getaddrinfo
Some of the variables that the user passes to the cppr::Request
class are passed directly to getaddrinfo. Therefore, it is
necessary to handle errors within the Getaddrinfo wrapper
function. Furthermore, debugging information is printed to the
console if connection-based error occurs, so the user may debug
their own program.

Add error handling and debugging messages to Getaddrinfo